STI with init_subclass
