+++ hydra_base/db/model/hydraconfig/__init__.py
@@ -1,6 +1,1 @@
-from .config_key_types import (
-    ConfigKey,
-    config_key_type_map
-)
-
 from .hydraconfig import *
